Do a bunch of renaming
