+++ modules/region_assignment/CommandAssignCN.js
@@ -3,7 +3,7 @@
 const
     config = require('config'),
 
-    CommandAssignmentBase = require('./command-assignment-base');
+    CommandAssignmentBase = require('./CommandAssignmentBase');
 
 class CommandAssignCN extends CommandAssignmentBase {
     constructor(module) {
