+++ modules/squads/CommandKick.js
@@ -3,7 +3,7 @@
 const
     config = require('config'),
 
-    CommandSquadBase = require('./command-squad-base'),
+    CommandSquadBase = require('./CommandSquadBase'),
     CommandError = require('../errors/CommandError');
 
 class CommandKick extends CommandSquadBase {
